+++ packages/components/radio/src/index.ts
@@ -2,4 +2,5 @@
 
 export { Radio, type RadioProps } from "./radio";
 export { RadioGroup, type RadioGroupProps } from "./radio-group";
+export { type RadioValue } from "./type";
 export { useRadio, type UseRadioProps } from "./use-radio";