+++ packages/components/rhf/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/rhf
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/radio@0.1.6
+
 ## 0.1.13
 
 ### Patch Changes