+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/react
 
+## 0.1.13
+
+### Patch Changes
+
+- feat: radio number value
+- Updated dependencies
+  - @jamsr-ui/radio@0.1.6
+
 ## 0.1.12
 
 ### Patch Changes